Move the molecular test charts to their own module.
